fix: preserve indexed record references
